+++ Modules/Intent.Modules.Application.MediatR/Templates/QueryInterface/QueryInterfaceTemplatePartial.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Intent.Engine;
 using Intent.Modules.Common;
+using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;
@@ -13,7 +14,7 @@ using Intent.Templates;
 namespace Intent.Modules.Application.MediatR.Templates.QueryInterface
 {
     [IntentManaged(Mode.Fully, Body = Mode.Merge)]
-    partial class QueryInterfaceTemplate : CSharpTemplateBase<object>
+    public partial class QueryInterfaceTemplate : CSharpTemplateBase<object>, ICSharpFileBuilderTemplate
     {
         [IntentManaged(Mode.Fully)]
         public const string TemplateId = "Intent.Application.MediatR.QueryInterface";
@@ -21,7 +22,12 @@ namespace Intent.Modules.Application.MediatR.Templates.QueryInterface
         [IntentManaged(Mode.Ignore, Signature = Mode.Fully)]
         public QueryInterfaceTemplate(IOutputTarget outputTarget, object model = null) : base(TemplateId, outputTarget, model)
         {
+            CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
+               .AddInterface("IQuery");
         }
+
+        [IntentManaged(Mode.Fully)]
+        public CSharpFile CSharpFile { get; }
 
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         protected override CSharpFileConfig DefineFileConfig()
@@ -30,6 +36,12 @@ namespace Intent.Modules.Application.MediatR.Templates.QueryInterface
                 className: $"IQuery",
                 @namespace: $"{this.GetNamespace()}",
                 relativeLocation: $"{this.GetFolderPath()}");
+        }
+
+        [IntentManaged(Mode.Fully)]
+        public override string TransformText()
+        {
+            return CSharpFile.ToString();
         }
     }
 }